refactor: reduced and reoganized code
