--- src/hooks/useTicket.js
@@ -1,5 +1,0 @@
-import { tickets } from '@/services/ticket'
-
-export default function useTicket() {
-  return { tickets }
-}
