+++ src/lib/react-query.js
@@ -1,0 +1,5 @@
+import { QueryClient } from '@tanstack/react-query'
+
+const queryClient = new QueryClient()
+
+export { queryClient }
