+++ src/main.jsx
@@ -3,9 +3,12 @@ import React from 'react'
 import ReactDOM from 'react-dom/client'
 import { RouterProvider } from 'react-router-dom'
 import router from './lib/router.jsx'
+import QueryContext from './contexts/QueryContext'
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
-    <RouterProvider router={router} />
+    <QueryContext>
+      <RouterProvider router={router} />
+    </QueryContext>
   </React.StrictMode>,
 )
