+++ src/services/ticket.js
@@ -1,6 +1,7 @@
 import axiosInstance from '@/lib/axios'
 import serverOptions from '@/config/server'
 import useAuth from '@/hooks/useAuth'
+import formatTicket from '@/utils/formatTicket'
 
 export async function fetchTickets() {
   return new Promise((resolve) => {
@@ -8,7 +9,10 @@ export async function fetchTickets() {
       .get(`unidades/1/painel?servicos=${serverOptions.services}`)
       .then(
         (response) => {
-          resolve(response.data)
+          const ticketResponse = response.data.map((ticket) =>
+            formatTicket(ticket),
+          )
+          resolve(ticketResponse)
         },
         (error) => {
           const originalRequest = error.config
@@ -29,26 +33,4 @@ export async function fetchTickets() {
         },
       )
   })
-}
-
-export async function tickets() {
-  const tickets = await fetchTickets()
-
-  const ticketsList = tickets.map((ticket) => normalizeMessage(ticket))
-
-  return await ticketsList
-}
-
-function normalizeMessage(data) {
-  return {
-    id: data.id,
-    type: 'ticket',
-    title: data.siglaSenha + ('000' + data.numeroSenha).slice(-3),
-    subtitle: data.local + ' ' + ('00' + data.numeroLocal).slice(-2),
-    description: data.prioridade,
-    paciente: data.nomeCliente,
-    guiche: data.numeroLocal,
-    setor: data.servico.nome,
-    $data: data,
-  }
 }
